// src/app/interface/decimo-terceiro.interface.ts
export interface DecimoTerceiroInterface {
    cpf: string;
    tipoSalario: string;
    dataAdmissao: string;
    parcela: string;
    salarioBase: number;
    jornada: number;
    dependentesIR: number;
    percenPensao: number;
    mediaSalarioVariavel: number;
    ano: number;
    ehSalarioPorHora: boolean;
}